working file extractor; removed hard-coded paths

+++ include/IDiscIO.hpp
@@ -4,6 +4,7 @@
 #include <memory>
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
 namespace NOD
 {
@@ -15,27 +16,27 @@ public:
 
     struct IReadStream
     {
-        virtual size_t read(void* buf, size_t length)=0;
-        virtual void seek(size_t offset, int whence=SEEK_SET)=0;
+        virtual uint64_t read(void* buf, uint64_t length)=0;
+        virtual void seek(int64_t offset, int whence=SEEK_SET)=0;
     };
-    virtual std::unique_ptr<IReadStream> beginReadStream(size_t offset=0) const=0;
+    virtual std::unique_ptr<IReadStream> beginReadStream(uint64_t offset=0) const=0;
 
     struct IWriteStream
     {
-        virtual size_t write(void* buf, size_t length)=0;
+        virtual uint64_t write(void* buf, uint64_t length)=0;
     };
-    virtual std::unique_ptr<IWriteStream> beginWriteStream(size_t offset=0) const=0;
+    virtual std::unique_ptr<IWriteStream> beginWriteStream(uint64_t offset=0) const=0;
 };
 
 struct IPartReadStream
 {
-    virtual void seek(size_t offset, int whence=SEEK_SET)=0;
-    virtual size_t read(void* buf, size_t length)=0;
+    virtual void seek(int64_t offset, int whence=SEEK_SET)=0;
+    virtual uint64_t read(void* buf, uint64_t length)=0;
 };
 
 struct IPartWriteStream
 {
-    virtual size_t write(void* buf, size_t length)=0;
+    virtual uint64_t write(void* buf, uint64_t length)=0;
 };
 
 }
